+++ src/Basescape/SoldierArmorState.cpp
@@ -42,7 +42,7 @@ namespace OpenXcom
  * @param base Pointer to the base to get info from.
  * @param soldier ID of the selected soldier.
  */
-SoldierArmorState::SoldierArmorState(Game *game, Base *base, unsigned int soldier) : State(game), _base(base), _soldier(soldier)
+SoldierArmorState::SoldierArmorState(Game *game, Base *base, size_t soldier) : State(game), _base(base), _soldier(soldier)
 {
 	_screen = false;
