+++ src/Interface/TextEdit.h
@@ -41,7 +41,7 @@ private:
 	bool _blink;
 	Timer *_timer;
 	wchar_t _ascii;
-	unsigned int _caretPos;
+	size_t _caretPos;
 
 	/// Checks if a character will exceed the maximum width.
 	bool exceedsMaxWidth(wchar_t c);
